Рефакторинг и комментирование кода
