+++ src/Serial/Series.cs
@@ -2,6 +2,9 @@
 
 namespace algorithms_cs.Serial;
 
+/// <summary>
+/// Series representing a sequence of increasing numbers over BufferedTapeReader.
+/// </summary>
 public class Series
 {
     private readonly BufferedTapeReader _lsTape;
@@ -13,6 +16,10 @@ public class Series
         _lsTape = tape;
     }
 
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <returns>A next value if it is greater than the previous value</returns>
     public SeriesReturn<double> Next()
     {
         var peekValue = _lsTape.Peek();
@@ -24,12 +31,8 @@ public class Series
                 _buffer = value.GetValue();
                 return new SeriesReturn<double>(value.GetValue());
             }
-            
             return new SeriesReturn<double>(SeriesReturnType.SeriesEnded);
-            
-            //return new CorrectSeriesReturn<double>(value.Value, SeriesReturnType.Correct);
         }
-
         return new SeriesReturn<double>(SeriesReturnType.TapeEnded);
     }
 }
