+++ src/Tape/Tape.cs
@@ -5,7 +5,14 @@
 /// </summary>
 public class Tape
 {
+    public readonly string? TapePath;
+    
     protected Tape()
     {
+    }
+    
+    protected Tape(string path)
+    {
+        TapePath = path;
     }
 }
